Fixed more room-related issues, added Play option for spectators
Lingering timer issues fixed.
Tabbing out during a game now correctly times out the player, even if a game had already finished.
Ranked games now continue to start after a game has finished.
FFA room should no longer close itself.
Split RoomManager into its own file.
