+++ public/index.html
@@ -274,29 +274,33 @@
             </table>
             <div class="status" id="statusArea">
                 <ul class="justify-flexbox vertical" id="roomManage">
-                    <li id="manageCpus">
+                    <li class="player" id="manageCpus">
                         <img class="roomManageIcon" src="images/mainpage/cpu.png">
                         <span>Manage CPUs</span>
                     </li>
-                    <li id="manageSettings">
+                    <li class="player" id="manageSettings">
                         <img class="roomManageIcon" src="images/mainpage/settings.png">
                         <span>Settings</span>
                     </li>
-                    <li id="manageRoomPassword">
+                    <li class="player" id="manageRoomPassword">
                         <img class="roomManageIcon" src="images/mainpage/settings.png">
                         <span>Set Room Password</span>
                     </li>
-                    <li id="manageJoinLink">
+                    <li class="player" id="manageJoinLink">
                         <img class="roomManageIcon" src="images/mainpage/link.png">
                         <span>Show Join Link</span>
                     </li>
-                    <li id="manageStartRoom">
+                    <li class="player" id="manageStartRoom">
                         <img class="roomManageIcon" src="images/mainpage/start.png">
                         <span>Start Room</span>
                     </li>
-                    <li id="manageSpectate">
+                    <li class="player" id="manageSpectate">
                         <img class="roomManageIcon" src="images/mainpage/spectate.png">
                         <span>Spectate</span>
+                    </li>
+                    <li id="managePlay">
+                        <img class="roomManageIcon" src="images/navbar/CustomIcon.png">
+                        <span>Play</span>
                     </li>
                 </ul>
                 <div id="statusMsg"></div>
